third_party/blink/web_tests/http/tests/cookies/single-quoted-value-expected.txt

Test for <https://bugs.webkit.org/show_bug.cgi?id=85775> [BlackBerry] Cookie parsing issue. If the cookie value provided was (") then the browser creates a session cookie instead.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that setting a cookie with a single-quoted value works.
PASS cookie is 'singleQuoted="value'.
PASS successfullyParsed is true

TEST COMPLETE
